relocate version to top-level utils
