+++ runway/utils/__init__.py
@@ -48,6 +48,7 @@ from ..compat import cached_property  # noqa pylint: disable=unused-import
 # make this importable without defining __all__ yet.
 # more things need to be moved of this file before starting an explicit __all__.
 from ._file_hash import FileHash  # noqa pylint: disable=unused-import
+from ._version import Version  # noqa pylint: disable=unused-import
 
 if TYPE_CHECKING:
     from mypy_boto3_cloudformation.type_defs import OutputTypeDef
