+++ runway/utils/_version.py
@@ -1,4 +1,4 @@
-"""Core utilities."""
+"""Version utilities."""
 from __future__ import annotations
 
 import packaging.version
