--- tests/unit/core/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Empty file for python import traversal."""
